extract module: runner::can_spawn
